requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
databricks-sdk
